+++ lib/service/Canvas.dart
@@ -59,7 +59,7 @@ class Canvas with Service {
     int? responseCode = response?.statusCode;
     String? responseString = response?.body;
     if (responseCode == 200) {
-      List<CanvasCourse>? courses = CanvasCourse.fromJsonList(JsonUtils.decodeList(responseString));
+      List<CanvasCourse>? courses = CanvasCourse.listFromJson(JsonUtils.decodeList(responseString));
       return courses;
     } else {
       Log.w('Failed to load canvas courses. Response:\n$responseCode: $responseString');
@@ -88,6 +88,40 @@ class Canvas with Service {
       return CanvasCourse.fromJson(courseJson);
     } else {
       Log.w('Failed to load canvas course. Response:\n$responseCode: $responseString');
+      return null;
+    }
+  }
+
+  Future<List<CanvasFile>?> loadFilesForCourse(int courseId) async {
+    if (!_available) {
+      return null;
+    }
+    String url = '${Config().canvasUrl}/api/v1/courses/$courseId/files';
+    http.Response? response = await Network().get(url, headers: _authHeaders);
+    int? responseCode = response?.statusCode;
+    String? responseString = response?.body;
+    if (responseCode == 200) {
+      List<CanvasFile>? files = CanvasFile.listFromJson(JsonUtils.decodeList(responseString));
+      return files;
+    } else {
+      Log.w('Failed to load canvas course files. Response:\n$responseCode: $responseString');
+      return null;
+    }
+  }
+
+  Future<List<CanvasFolder>?> loadFoldersForCourse(int courseId) async {
+    if (!_available) {
+      return null;
+    }
+    String url = '${Config().canvasUrl}/api/v1/courses/$courseId/folders';
+    http.Response? response = await Network().get(url, headers: _authHeaders);
+    int? responseCode = response?.statusCode;
+    String? responseString = response?.body;
+    if (responseCode == 200) {
+      List<CanvasFolder>? folders = CanvasFolder.listFromJson(JsonUtils.decodeList(responseString));
+      return folders;
+    } else {
+      Log.w('Failed to load canvas course folders. Response:\n$responseCode: $responseString');
       return null;
     }
   }